+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased Changes
 
 * Feature - TODO
 
+* Issue - Falls back to header request checksums when using custom endpoints or endpoint providers for PutObject and UploadPart operations.
+
 1.210.1 (2026-01-06)
 ------------------